Meta information change to reflect actual application use
Didn’t change the meta so when I copied and pasted the usual
information, I didn’t change out the previous app’s meta tags. This has
been fixed and now the head should show that this app is for the
keystroke calculator and not the mood of the words.